% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpSpatialDist3D.R
\name{PipeOpSpatialDist3D}
\alias{PipeOpSpatialDist3D}
\title{Add buffer distances to 3D reference locations as features}
\value{
a `PipeOpSpatialDist3D` R6 class object
}
\description{
Adds new variables based on the 3D Euclidean distances to reference locations.
}
\examples{
library(mlr3)
library(mlr3pipelines)

task = tsk('boston_housing')
pop = PipeOpSpatialDist3D$new()
pop$param_set$values$lat = 'lat'
pop$param_set$values$lon = 'lon'
pop$train(list(task))
pop$predict(list(task))[[1]]$data()
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpSpatialDist3D}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpSpatialDist3D-new}{\code{PipeOpSpatialDist3D$new()}}
\item \href{#method-PipeOpSpatialDist3D-clone}{\code{PipeOpSpatialDist3D$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpSpatialDist3D-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpSpatialDist3D-new}{}}}
\subsection{Method \code{new()}}{
Create a new PipeOpSpatialDist3D object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDist3D$new(
  id = "geodist3d",
  param_vals = list(prefix = "geodist", minimum = TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, identifier for the class instance.}

\item{\code{param_vals}}{named list of hyperparameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PipeOpSpatialDist3D` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpSpatialDist3D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpSpatialDist3D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDist3D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
