% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpSpatialDistCluster.R
\name{PipeOpSpatialDistCluster}
\alias{PipeOpSpatialDistCluster}
\title{Add 2D Euclidean distances to cluster centroids as features}
\value{
a `PipeOpSpatialDistCluster` object.
}
\description{
Adds new variables based on the 2D Euclidean distances to cluster centroids
in the data. These centroids are generated by k-means clustering of the
variables provided in 'lat' and 'lon' as well as the target variable.
}
\examples{
library(mlr3)
library(mlr3pipelines)

task = tsk('boston_housing')
pop = PipeOpSpatialDistCluster$new()
pop$param_set$values$lat = 'lat'
pop$param_set$values$lon = 'lon'
pop$train(list(task))
pop$predict(list(task))[[1]]$data()
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpSpatialDistCluster}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpSpatialDistCluster-new}{\code{PipeOpSpatialDistCluster$new()}}
\item \href{#method-PipeOpSpatialDistCluster-clone}{\code{PipeOpSpatialDistCluster$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpSpatialDistCluster-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpSpatialDistCluster-new}{}}}
\subsection{Method \code{new()}}{
Create a new PipeOpSpatialDistCluster object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDistCluster$new(
  id = "spatialdist_cluster",
  param_vals = list(prefix = "geodist", k = 5)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, identifier for the class instance.}

\item{\code{param_vals}}{named list of hyperparameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PipeOpSpatialDistCluster` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpSpatialDistCluster-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpSpatialDistCluster-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDistCluster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
