% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpSpatialDistCluster3D.R
\name{PipeOpSpatialDistCluster3D}
\alias{PipeOpSpatialDistCluster3D}
\title{Add 3D Euclidean distances to cluster centroids as features}
\value{
a `PipeOpSpatialDistCluster3D` object.
}
\description{
Adds new variables based on the 3D Euclidean distances to cluster centroids
in the data. These centroids are generated by k-means clustering of the
variables provided in 'lat', 'lon', and 'depth', as well as the target
variable.
}
\section{Super classes}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{\link[mlr3pipelines:PipeOpTaskPreproc]{mlr3pipelines::PipeOpTaskPreproc}} -> \code{\link[mlr3pipelines:PipeOpTaskPreprocSimple]{mlr3pipelines::PipeOpTaskPreprocSimple}} -> \code{PipeOpGeodist3DCluster3D}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipeOpGeodist3DCluster3D-new}{\code{PipeOpSpatialDistCluster3D$new()}}
\item \href{#method-PipeOpGeodist3DCluster3D-clone}{\code{PipeOpSpatialDistCluster3D$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="help"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-help'><code>mlr3pipelines::PipeOp$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-predict'><code>mlr3pipelines::PipeOp$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-print'><code>mlr3pipelines::PipeOp$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train"><a href='../../mlr3pipelines/html/PipeOp.html#method-PipeOp-train'><code>mlr3pipelines::PipeOp$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpGeodist3DCluster3D-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpGeodist3DCluster3D-new}{}}}
\subsection{Method \code{new()}}{
Create a new PipeOpSpatialDistCluster3D object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDistCluster3D$new(
  id = "geodist_3dcluster",
  param_vals = list(k = 10, prefix = "geodist")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, identifier for the class instance.}

\item{\code{param_vals}}{named list of hyperparameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PipeOpSpatialDistCluster3D` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipeOpGeodist3DCluster3D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipeOpGeodist3DCluster3D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpSpatialDistCluster3D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
