% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3spatialops-package.R
\docType{package}
\name{mlr3spatialops-package}
\alias{mlr3spatialops}
\alias{mlr3spatialops-package}
\title{mlr3spatialops: Spatial Pipeline Operations for 'mlr3pipelines'}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
